Align constants
