+++ system-parachains/bulletin/bulletin-polkadot/src/genesis_config_presets.rs
@@ -20,7 +20,10 @@ use alloc::{vec, vec::Vec};
 use cumulus_primitives_core::ParaId;
 use frame_support::build_struct_json_patch;
 use parachains_common::{AccountId, AuraId};
-use polkadot_constants::{currency::UNITS as DOT, SAFE_XCM_VERSION};
+use system_parachains_constants::{
+	genesis_presets::SAFE_XCM_VERSION,
+	polkadot::currency::UNITS as DOT,
+};
 use sp_core::sr25519;
 use sp_genesis_builder::PresetId;
 use sp_keyring::Sr25519Keyring;
