+++ system-parachains/bulletin/bulletin-polkadot/src/lib.rs
@@ -24,7 +24,6 @@ include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 mod apis;
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarks;
-pub mod constants;
 mod genesis_config_presets;
 pub mod migrations;
 mod weights;
@@ -53,7 +52,18 @@ use parachains_common::{
 	impls::DealWithFees,
 	message_queue::{NarrowOriginToSibling, ParaIdToSibling},
 	AccountId, AuraId, Balance, BlockNumber, Hash, Header, Nonce, Signature,
-	AVERAGE_ON_INITIALIZE_RATIO,
+};
+use system_parachains_constants::{
+	async_backing::{
+		AVERAGE_ON_INITIALIZE_RATIO, HOURS, MAXIMUM_BLOCK_WEIGHT, MILLISECS_PER_BLOCK,
+	},
+	polkadot::{
+		consensus::*,
+		currency::{
+			system_para_deposit as deposit, CENTS, MILLICENTS, SYSTEM_PARA_EXISTENTIAL_DEPOSIT,
+			UNITS,
+		},
+	},
 };
 use polkadot_runtime_common::{BlockHashCount, SlowAdjustingFeeUpdate};
 use sp_core::{crypto::KeyTypeId, OpaqueMetadata};
@@ -78,16 +88,62 @@ use xcm_runtime_apis::{
 	fees::Error as XcmPaymentApiError,
 };
 
-pub use constants::polkadot_constants;
+/// Bulletin uses 6s slot duration (same as block time).
+pub const SLOT_DURATION: u64 = MILLISECS_PER_BLOCK;
 
-use polkadot_constants::{
-	consensus::*,
-	currency::{deposit, CENTS, EXISTENTIAL_DEPOSIT, MILLICENTS},
-	fee::WeightToFee,
-	time::*,
-};
-// Re-export for benchmarks
-pub use polkadot_constants::currency::UNITS;
+/// Handles converting a weight scalar to a fee value, based on the scale and granularity
+/// of the node's balance type.
+///
+/// Takes the maximum of ref_time and proof_size fees.
+pub struct WeightToFee;
+impl frame_support::weights::WeightToFee for WeightToFee {
+	type Balance = Balance;
+
+	fn weight_to_fee(weight: &Weight) -> Self::Balance {
+		use frame_support::weights::{
+			FeePolynomial, WeightToFeeCoefficient, WeightToFeeCoefficients, WeightToFeePolynomial,
+		};
+		use smallvec::smallvec;
+
+		/// Maps the reference time component of `Weight` to a fee.
+		struct RefTimeToFee;
+		impl WeightToFeePolynomial for RefTimeToFee {
+			type Balance = Balance;
+			fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
+				let p = CENTS;
+				let q = 200 * Balance::from(ExtrinsicBaseWeight::get().ref_time());
+				smallvec![WeightToFeeCoefficient {
+					degree: 1,
+					negative: false,
+					coeff_frac: Perbill::from_rational(p % q, q),
+					coeff_integer: p / q,
+				}]
+			}
+		}
+
+		/// Maps the proof size component of `Weight` to a fee.
+		struct ProofSizeToFee;
+		impl WeightToFeePolynomial for ProofSizeToFee {
+			type Balance = Balance;
+			fn polynomial() -> WeightToFeeCoefficients<Self::Balance> {
+				let p = CENTS;
+				let q = 20_000;
+				smallvec![WeightToFeeCoefficient {
+					degree: 1,
+					negative: false,
+					coeff_frac: Perbill::from_rational(p % q, q),
+					coeff_integer: p / q,
+				}]
+			}
+		}
+
+		let time_poly: FeePolynomial<Balance> = RefTimeToFee::polynomial().into();
+		let proof_poly: FeePolynomial<Balance> = ProofSizeToFee::polynomial().into();
+
+		// Take the maximum instead of the sum to charge by the more scarce resource.
+		time_poly.eval(weight.ref_time()).max(proof_poly.eval(weight.proof_size()))
+	}
+}
 
 /// The address format for describing accounts.
 pub type Address = MultiAddress<AccountId, ()>;
@@ -252,7 +308,7 @@ impl pallet_authorship::Config for Runtime {
 }
 
 parameter_types! {
-	pub const ExistentialDeposit: Balance = EXISTENTIAL_DEPOSIT;
+	pub const ExistentialDeposit: Balance = SYSTEM_PARA_EXISTENTIAL_DEPOSIT;
 	pub const RandomParaId: ParaId = ParaId::new(43211234);
 }
